BS-262 | Fix. Remove concrete dependency on ValueSetTask
